+++ ron-rdt/lib/RON/Data/RGA.hs
@@ -12,7 +12,7 @@
 -- | Replicated Growable Array (RGA)
 module RON.Data.RGA
     ( RGA (..)
-    , RgaRaw
+    , RgaRep
     , RgaString
     , edit
     , editText
@@ -111,7 +111,7 @@ vertexListFromOps = foldr go mempty where
                 HashMap.insert opId (item $ Just listHead) listItems
 
 -- | Untyped RGA
-newtype RgaRaw = RgaRaw (Maybe VertexList)
+newtype RgaRep = RgaRep (Maybe VertexList)
     deriving (Eq, Monoid, Semigroup, Show)
 
 data PatchSet = PatchSet
@@ -164,12 +164,12 @@ patchSetFromChunk ReducedChunk{rcRef, rcBody} =
                 Just patch -> mempty{psPatches = Map.singleton rcRef patch}
                 Nothing -> mempty
 
-instance Reducible RgaRaw where
+instance Reducible RgaRep where
     reducibleOpType = rgaType
 
-    stateFromChunk = RgaRaw . vertexListFromOps
+    stateFromChunk = RgaRep . vertexListFromOps
 
-    stateToChunk (RgaRaw rga) = StateChunk
+    stateToChunk (RgaRep rga) = StateChunk
         {stateType = rgaType, stateVersion = chunkVersion stateBody, stateBody}
       where
         stateBody = maybe [] vertexListToOps rga
@@ -203,7 +203,7 @@ reapplyPatchSet :: PatchSet -> PatchSet
 reapplyPatchSet ps =
     continue ps [reapplyPatchesToOtherPatches, reapplyRemovalsToPatches]
 
-reapplyPatchSetToState :: RgaRaw -> PatchSet -> (RgaRaw, PatchSet)
+reapplyPatchSetToState :: RgaRep -> PatchSet -> (RgaRep, PatchSet)
 reapplyPatchSetToState rga ps =
     continue (rga, ps) [reapplyPatchesToState, reapplyRemovalsToState]
 
@@ -212,13 +212,13 @@ continue x fs = case asum $ map ($ x) fs of
     Nothing -> x
     Just x' -> continue x' fs
 
-reapplyPatchesToState :: (RgaRaw, PatchSet) -> Maybe (RgaRaw, PatchSet)
-reapplyPatchesToState (RgaRaw rstate, ps@PatchSet{..}) = case rstate of
+reapplyPatchesToState :: (RgaRep, PatchSet) -> Maybe (RgaRep, PatchSet)
+reapplyPatchesToState (RgaRep rstate, ps@PatchSet{..}) = case rstate of
     Just VertexList{listHead = targetHead, listItems = targetItems} -> asum
         [ do
             targetItems' <- applyPatch parent patch targetItems
             pure
-                ( RgaRaw . Just $ VertexList targetHead targetItems'
+                ( RgaRep . Just $ VertexList targetHead targetItems'
                 , ps{psPatches = Map.delete parent psPatches}
                 )
         | (parent, patch) <- Map.assocs psPatches
@@ -227,7 +227,7 @@ reapplyPatchesToState (RgaRaw rstate, ps@PatchSet{..}) = case rstate of
         -- rstate is empty => only virtual 0 node exists
         -- => we can apply only 0 patch
         patch <- psPatches !? Zero
-        pure (RgaRaw $ Just patch, ps{psPatches = Map.delete Zero psPatches})
+        pure (RgaRep $ Just patch, ps{psPatches = Map.delete Zero psPatches})
 
 reapplyPatchesToOtherPatches :: PatchSet -> Maybe PatchSet
 reapplyPatchesToOtherPatches ps@PatchSet{..} = asum
@@ -260,14 +260,14 @@ applyPatch parent patch targetItems = case parent of
         let item' = item{itemNext = Just next'}
         pure $ HashMap.insert parent item' targetItems <> newItems
 
-reapplyRemovalsToState :: (RgaRaw, PatchSet) -> Maybe (RgaRaw, PatchSet)
-reapplyRemovalsToState (RgaRaw rstate, ps@PatchSet{..}) = do
+reapplyRemovalsToState :: (RgaRep, PatchSet) -> Maybe (RgaRep, PatchSet)
+reapplyRemovalsToState (RgaRep rstate, ps@PatchSet{..}) = do
     VertexList{listHead = targetHead, listItems = targetItems} <- rstate
     asum
         [ do
             targetItems' <- applyRemoval parent tombstone targetItems
             pure
-                ( RgaRaw . Just $ VertexList targetHead targetItems'
+                ( RgaRep . Just $ VertexList targetHead targetItems'
                 , ps{psRemovals = Map.delete parent psRemovals}
                 )
         | (parent, tombstone) <- Map.assocs psRemovals